Create BasePillow UI Component
