+++ packages/ui/package.json
@@ -53,6 +53,7 @@
     "@emotion/react": "11.13.3",
     "@emotion/styled": "11.13.0",
     "@radix-ui/react-dialog": "1.1.1",
+    "@radix-ui/react-slot": "1.1.0",
     "@radix-ui/react-tabs": "1.1.0",
     "@vanilla-extract/css": "1.15.5",
     "@vanilla-extract/dynamic": "2.1.2",
