+++ packages/ui/src/index.ts
@@ -1,4 +1,5 @@
 export { Alert } from './components/Alert/Alert'
+export { BasePillow, pillowSize } from './components/BasePillow/BasePillow'
 export { HedvigLogo } from './components/HedvigLogo/HedvigLogo'
 export { HedvigSymbol } from './components/HedvigSymbol/HedvigSymbol'
 export { Space } from './components/Space'
